Moved docstrings to own module
Created an official class to subclass docstrings from as well
