+++ botocore/docs/method.py
@@ -16,7 +16,6 @@ from botocore.docs.params import RequestParamsDocumenter
 from botocore.docs.params import ResponseParamsDocumenter
 from botocore.docs.example import ResponseExampleDocumenter
 from botocore.docs.example import RequestExampleDocumenter
-from botocore.docs.bcdoc.restdoc import DocumentStructure
 
 
 def get_instance_public_methods(instance):
@@ -236,62 +235,3 @@ def document_model_driven_method(section, method_name, operation_model,
                 include=include_output, exclude=exclude_output)
     else:
         return_section.write(':returns: None')
-
-
-class LazyLoadedDocstring(str):
-    """Used for lazily loading docstrings
-
-    You can instantiate this class and assign it to a __doc__ value.
-    The docstring will not be generated till accessed via __doc__ or
-    help().
-    """
-    def __init__(self, *args, **kwargs):
-        """
-        The args and kwargs are the same as the underlying document
-        generation function. These just get proxied to the underlying
-        function.
-        """
-        super(LazyLoadedDocstring, self).__init__()
-        self._gen_args = args
-        self._gen_kwargs = kwargs
-        self._docstring = None
-        self._docstring_writer = document_model_driven_method
-
-    def __new__(cls, *args, **kwargs):
-        # Needed in order to sub class from str with args and kwargs
-        return super(LazyLoadedDocstring, cls).__new__(cls)
-
-    def expandtabs(self, tabsize=8):
-        """Expands tabs to spaces
-
-        So this is a big hack in order to get lazy loaded docstring work
-        for the ``help()``. In the ``help()`` function, ``pydoc`` and
-        ``inspect`` are used. At some point the ``inspect.cleandoc``
-        method is called. To clean the docs ``expandtabs`` is called
-        and that is where we override the method to generate and return the
-        docstrings.
-        """
-        if self._docstring is None:
-            self._generate().expandtabs(tabsize)
-        return self._docstring.expandtabs(tabsize)
-
-    def __str__(self):
-        return self._generate()
-
-    # __doc__ of target will use either __repr__ or __str__ of this class.
-    __repr__ = __str__
-
-    def _generate(self):
-        # Generate the docstring if it is not already cached.
-        if self._docstring is None:
-            self._docstring = self._create_docstring()
-        return self._docstring
-
-    def _create_docstring(self):
-        docstring_structure = DocumentStructure('docstring')
-        # Call the document method function with the args and kwargs
-        # passed to the class.
-        self._docstring_writer(
-            docstring_structure, *self._gen_args,
-            **self._gen_kwargs)
-        return docstring_structure.flush_structure().decode('utf-8')
